Refactor
1. WheelerPlayerController

a. added SetParticleSystemElementType fn.  NOTE: This probably requires the particle system to be set up in advance, to use custom data, and a specific kind

b. ellietodo: make sure we set up the custom data thing in code, this fn is the place for it

2. ScannableObject

a. added GetParticleElementType
b. removed some unnecessary vars