Updated SpriteVariable and SpriteReference
1. SpriteVariable now inherits from Scriptable Object, and simply duplicates the functionality in GenericVariable<T>.

I couldn't think of a way to attach the PreviewField to the value of the class, b/c it has to apply, only to GenericVariable<Sprite>, and it doesn't want to be attached to something like GenericVariable<float>.

Since there's very little in the GenericVariable<T> class anyway, I just duplicated it.  If nothing else, it will save a lot of time, compared to trying to figure out the right way to do it.  My suspicion is, OdinInspector doesn't support the kind of member-hiding override we're trying to use in this class, and that we'd have to work around their system anyway

2.